requirements.txt
urllib3==1.25.9
Django==3.2.9
djangorestframework==3.13.1
PyJWT==2.3.0
